Moved token generation/validation code to new helper module
